+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 
 **IMPORTANT 2**: All CSS variables have been renamed. You will need to update any custom CSS that use them.
 
-**IMPORTANT 3**: Your current `{{history}}` macro may no longer work as it did before. You will need to reconfigure it in the new interface by clicking "Edit" under "Summarization"
+**IMPORTANT 3**: The short-term and long-term memory macros have been renamed. You will need to update anywhere you use them.
+
+**IMPORTANT 4**: Your current `{{history}}` macro may no longer work as it did before. You will need to reconfigure it in the new interface by clicking "Edit" under "Summarization"
 
 - **New Feature**: Reworked "Edit" interface for the summary prompt. This is now where the `{{history}}`, `{{message}}`, and `{{words}}` macro are defined and modified. You can also create custom macros for the summary prompt using either a message range or STScript.
 - **New Feature**: A new setting called `Static Memory Mode` can be enabled in the general injection settings. This mode makes long-term memories always be injected separately from short-term memories, regardless of context. Disabled by default.
@@ -14,6 +16,7 @@
 - **New Slash Command**: `/qm-set` allows you to set the memory for a message.
 - **Changed**: All slash commands now use "-" instead of "_" and start with "qm-" (with "qvink-memory-" as an alias) to avoid overlapping with other commands. See the README for the updated list of all commands.
 - **Changed**: All CSS variables now use "-" instead of "_", and are prefixed with "--qm-"
+- **Changed**: The memory macros have been renamed to `{{qm-short-term-memory}}` and `{{qm-long-term-memory}}`.
 - **Changed**: Messages that have been removed from context are now turned grey in the chat. This can be modified by using the `--qm-message-removed` CSS variable.
 - **Changed**: Summaries that aren't included in memory can now be changed style with the `--qm-default` CSS variable.
 - **Changed**: The slash command `/qm-get` (previously `/get_memory`) can now accept a range of indexes and a custom separator.